feat: error handlers 503 and 500 are placed in the fastapi application exception_handler
