+++ services/ml_models/app/api/exceptions.py
@@ -1,0 +1,50 @@
+from fastapi import FastAPI, Request
+from fastapi.responses import JSONResponse
+from psycopg2 import OperationalError, InterfaceError, IntegrityError, DatabaseError
+
+from app.logs import get_logger
+
+logger = get_logger(__name__)
+
+def setup_exception_handlers(app: FastAPI):
+    
+    logger.info("Настройка обработчиков ошибок...")
+
+    @app.exception_handler(404)
+    async def not_found_handler(request: Request, exc: Exception):
+        logger.warning(f"URL не найден: {request.method} {request.url.path}")
+        return JSONResponse(
+            status_code=404,
+            content={"detail": f"Endpoint {request.method} {request.url.path} не найден"}
+        )
+    logger.info(" ✅ Обработчик 404 (Not Found)")
+
+    @app.exception_handler(OperationalError)
+    async def operational_error_handler(request: Request, exc: OperationalError):
+        logger.error(f"Ошибка подключения к БД (OperationalError): {exc}", exc_info=True)
+        return JSONResponse(
+            status_code=503,
+            content={"detail": "Ошибка подключения к базе данных."}
+        )
+    logger.info(" ✅ Обработчик 503 (OperationalError)")
+
+    @app.exception_handler(InterfaceError)
+    async def interface_error_handler(request: Request, exc: InterfaceError):
+        logger.error(f"Ошибка подключения к БД (InterfaceError): {exc}", exc_info=True)
+        return JSONResponse(
+            status_code=503,
+            content={"detail": "Ошибка подключения к базе данных."}
+        )
+    logger.info(" ✅ Обработчик 503 (InterfaceError)")
+
+    @app.exception_handler(Exception)
+    async def global_exception_handler(request: Request, exc: Exception):
+        logger.error(f"Глобальная неизвестная ошибка: {exc}", exc_info=True)
+        return JSONResponse(
+            status_code=500,
+            content={"detail": "Неизвестная внутренняя ошибка сервера."}
+        )
+    logger.info(" ✅ 500 ERROR")
+    
+
+    logger.info("✅ Добавлены обработчики ошибок")
