+++ services/ml_models/main.py
@@ -1,9 +1,9 @@
 import uvicorn
-from fastapi import FastAPI, Request
-from fastapi.responses import JSONResponse
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 
 from app.core.health import check_health_all
+from app.api.exceptions import setup_exception_handlers
 from app.api.routers import routers
 from app.logs import get_logger
 
@@ -20,6 +20,7 @@ app = FastAPI(
 
 logger.info("✅ app создано")
 
+setup_exception_handlers(app)
 app.include_router(routers)
 
 logger.info("✅ Добавлены эндпоинты")
@@ -31,16 +32,6 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
-@app.exception_handler(Exception)
-async def global_exception_handler(request: Request, exc: Exception):
-    logger.error(f"Глобальная неизвестная ошибка: {exc}", exc_info=True)
-    return JSONResponse(
-        status_code=500,
-        content={"detail": "Неизвестная внутренняя ошибка сервера."}
-    )
-
-logger.info("✅ Добавлены обработчики ошибок")
 
 if __name__ == "__main__":
     uvicorn.run(
